+++ src/games/scenarios/ALPGameScenario.java
@@ -24,7 +24,7 @@ public class ALPGameScenario extends GameScenario {
 	}
 
 	@Override
-	public double play(BoardGame game) {
+	public int play(BoardGame game) {
 		while (!game.endOfGame()) {
 			List<? extends GameMove> moves = game.findMoves();
 			if (!moves.isEmpty()) {
