+++ src/games/scenarios/RandomizedTwoPlayersGameScenario.java
@@ -21,7 +21,7 @@ public class RandomizedTwoPlayersGameScenario extends GameScenario {
 	}
 
 	@Override
-	public double play(BoardGame game) {
+	public int play(BoardGame game) {
 		while (!game.endOfGame()) {
 			List<? extends GameMove> moves = game.findMoves();
 			if (!moves.isEmpty()) {
